Move some helpers into AbstractAlgebra.PrettyPrinting
These are useful for writing tests also for downstream packages
